refactor(live): decompose LiveTradingEngine.__init__ into phase helpers (#486)
Step A of the live-engine modularization: the ~534-line __init__ monolith is
now a thin ~110-line orchestrator that delegates to 15 cohesive, single-purpose
private initializer helpers (_validate_inputs, _resolve_settings,
_init_coordinators, _init_risk_manager, _init_risk_policies,
_init_partial_operations, _init_correlation, _init_database,
_init_dynamic_risk_manager, _init_exchange_interface,
_resume_balance_from_snapshot, _init_strategy_manager, _seed_trading_state,
_init_time_exit_policy, _install_signal_handlers).

Every block moved verbatim — construction ordering, the full 35-param public
constructor signature, and every public attribute are preserved exactly.

Splitting the assignments across methods removed mypy's same-__init__ deferral
and surfaced a latent type inconsistency: LiveLoopTimingEngineState declared
data_freshness_threshold as float while the engine attribute, MarketDataHandler,
and the sibling interval fields are all int. Aligned the Protocol to int.

Pure refactor — backtest determinism fingerprint byte-identical
(sha256=ee76fd681362f5a251f9bd34ee40c7177ca8697e9b29e70b2c0d1ed2afd03a87,
trades=14, final_balance=9964.469867425983). Full fast suite green (4094
passed); mypy/ruff/black/bandit clean.

+++ src/engines/live/loop_timing.py
@@ -36,7 +36,7 @@ class LiveLoopTimingEngineState(Protocol):
     base_check_interval: int
     min_check_interval: int
     max_check_interval: int
-    data_freshness_threshold: float
+    data_freshness_threshold: int
     live_position_tracker: LivePositionTracker
     _ws_kline_active: bool
     _kline_buffer: KlineBuffer | None
